Moving gltf workers to separate extension
