+++ metacitygl/extensions/gtlf/worker.ts
@@ -1,4 +1,4 @@
-import * as Utils from "../../../utils";
+import * as Utils from "../../utils";
 
 declare var self: any;
